[ISSUE-192} carvaan-wallets types have been shifted to new file fr better code readability
