+++ packages/caravan-wallets/src/index.ts
@@ -1,7 +1,3 @@
-// eslint-disable-next-line @typescript-eslint/ban-ts-comment
-// @ts-ignore
-import { version } from "../package.json";
-import { UNSUPPORTED, UnsupportedInteraction } from "./interaction";
 import {
   BITBOX,
   BitBoxGetMetadata,
@@ -61,44 +57,11 @@ import {
   convertLegacyInput,
   convertLegacyOutput,
   getUnsignedMultisigPsbtV0,
-} from "@caravan/psbt";
+} from "@caravan/psbt"; 
 
-/**
- * Current @caravan/wallets version.
- */
-export const VERSION: string = version;
+import { KEYSTORE_TYPES } from "./type";
+import { UNSUPPORTED, UnsupportedInteraction } from "./interaction";
 
-export const MULTISIG_ROOT = "m/45'";
-
-/**
- * Keystores which support direct interactions.
- */
-export const DIRECT_KEYSTORES = {
-  BITBOX,
-  TREZOR,
-  LEDGER,
-  LEDGER_V2,
-} as const;
-
-/**
- * Keystores which support indirect interactions.
- */
-export const INDIRECT_KEYSTORES = {
-  HERMIT,
-  COLDCARD,
-  CUSTOM,
-} as const;
-
-/**
- * Supported keystores.
- */
-export const KEYSTORES = {
-  ...DIRECT_KEYSTORES,
-  ...INDIRECT_KEYSTORES,
-} as const;
-
-type KEYSTORE_KEYS = keyof typeof KEYSTORES;
-export type KEYSTORE_TYPES = (typeof KEYSTORES)[KEYSTORE_KEYS];
 
 /**
  * Return an interaction class for obtaining metadata from the given
@@ -643,6 +606,7 @@ export function ConfigAdapter({
   }
 }
 
+
 export * from "./interaction";
 export * from "./bitbox";
 export * from "./bcur";
@@ -653,3 +617,5 @@ export * from "./ledger";
 export * from "./trezor";
 export * from "./policy";
 export * from "./types";
+export * from "./type";
+
